+++ src/components/sections/layout/SectionTitle.tsx
